Open Taipei Pass login page and enter login info
At the beginning, I use the detach experimental option of ChromeOptions
to prevent window closing immediately and set window to maximum size.

In every page, I use WebDriverWait with expected_conditions module to
wait for specific element, which we want to interact with, to be
clickable. In this way, we can avoid "element is not found" error.

I use a json file to store login information, and extract the data when
I need to enter account and password. This file is only stored in local
directory.

Create .gitignore file to add login_info.json file, in case I
push this file to GitHub.